Main logic into "dir_merge" and "dir_merge_runner"
Rewrote main (now "dir_merge") to handle cmd line arguments only. If
arguments are given, then those directories are indexed directly.
Otherwise the repl will request input directories (this functionality
still needs to be implemented).
