+++ config.py
@@ -1,5 +1,7 @@
 from pathlib import Path
 
+# C:/Users/trjji/Documents/Coding Projects/obsidian-helpers/testing/test_dirs/test_dir_one
+# C:/Users/trjji/Documents/Coding Projects/obsidian-helpers/testing/test_dirs/test_dir_two
 
 DEFAULT_PATH_A = Path("testing/real_dirs/temp_clone_local")
 DEFAULT_PATH_B = Path("testing/real_dirs/temp_clone_remote")
